Keep Shiki off the homepage by splitting blog slug parsing.
The app shell only needs getSlugFromPathname to classify chrome. Importing
blog-post.tsx pulled markdown highlighting onto every `/` document.
